Create and implement basic functionality for Window class
The class allows for the construction of a GUI, although there is currently
no content to display. When the `wait_for_close` method is called on a
Window object, the GUI continuously redraws itself. Furthermore, the GUI
will properly terminate when the user closes it.

In future updates, the class will be expanded to include the ability to
display the maze and demonstrate the maze solver in action.