Add functionality for drawing cells on the window
Introduce the `draw_cell` method to the Window class, enabling the
drawing of cells on the window's canvas. This method takes a Cell object
and a color as parameters. When called, it draws the Cell object onto
the canvas with the specified color.

Additionally, a new class called Cell has been implemented. The Cell
class takes a Line object as a parameter in its constructor, storing the
coordinates of its points in four data members. The draw method utilizes
these coordinates to construct four separate lines, representing the
walls of the cell.

The Cell class also includes four additional data members that determine
the presence of walls within the cell. These values play a role in
determining which adjacent cell the maze solver will be allowed to move
to.

These additions expand the capabilities of the Window class, allowing
for the representation of cells in the GUI.