Add functionality for randomized maze generation
Introuduce A new method, __create_pathways() to generate pathways between
cells in a randomized manner.

The __create_pathways() method takes into account the current cell's
location (i and j) and the visited status of the cell. By iteratively
selecting a random direction from the available unvisited neighboring
cells, the method establishes connections by removing walls between the
current cell and the next cell.

The __create_entrance_and_exit() method is also introduced to
remove the top wall of the first cell and the bottom wall of the last
cell, creating an entrance and exit point for the maze.

The Maze class is also given a new data member called seed to provide
the random number genearor consistant results for testing purposes.