Implement maze solving, and visualization
- Create the Movement class to represent a movement between two cells in
the maze. Each Movement instance holds references to the starting cell
(cell_1) and the destination cell (cell_2).

- Create the MazeSolver class, which takes a maze as input and solves it
using a depth first exploring algorithm. The solve() method initiates
the solving process, and the private __solve_r() method recursively
explores the maze, keeping track of visited cells and movements made.
The solved route is stored in the route attribute of the MazeSolver
instance.

- Create the main() function. It creates a Window instance for visualization
and initializes a Maze object with specified parameters. The MazeSolver
class is used to solve the maze, and the resulting maze and solved route
are drawn on the window canvas using methods from the Window class.