Add InputHandler class for keyboard input handling
This commit lays the groundwork for turning the maze solver into a
game where the player races the computer to complete their maze first.

The newly introduced InputHandler class enables keyboard input handling
within the application. It serves as a bridge between the user's
interaction with the keyboard and the game's mechanics. It uses
dependencies from the "pynput.keyboard" module to achieve this.

During initialization, the InputHandler class receives a window object,
representing the application window. This allows the class to integrate
with the graphical user interface, ensuring the InputHandler class is
capeable of drawing user movements. Additionally, a controller object is
created, providing the necessary functionality to control keyboard input.