Update visuals to use images for players
I have made several updates to enhance the game's visual experience. The
computer and player now have dedicated images to represent them. The
computer is represented by an image of a computer, while the player is
represented by a heart image to symbolize human players.

Additionally, both the computer and player are now displayed on the same
maze, creating a more cohesive gameplay environment.

To accommodate the new images, the Window class's `draw_move` method was
replaced with a `move_img` method. This change allows for smoother
integration of the images. Also, trailing lines were removed to prevent
conflicts between the computer and player lines on the shared maze.

Furthermore, I renamed the InputHandler class to Human to better reflect
its purpose of managing the human player. To handle the computer player's
responsibilities, I created a new class called Computer. The Maze's `solve`
method was moved to the Computer class, as it aligns more closely with
the responsibilities of the Computer class. Notably, both the Player and
Computer classes now store references to their respective images.

These updates aim to create a more immersive gaming experience and
improve the overall visual quality of the game.