Terminate game functions when player reaches maze end
I introduce the capability to terminate all game functions when a player
reaches the end of a maze. The implementation of this functionality lays
the foundation for displaying a results screen in the future.

To achieve this, I have refactored the codebase to enhance player control.
Both the Human and Computer classes now store references to images, with
the Human class additionally storing coordinate information. The
responsibility for actual movement, however, has been shifted to the new
PlayerManager class. This was necessary to establish communication between
the start methods, so that they can notify each other to stop, by utilizing
the terminate flag, when the win condition is achieved.

Overall this helps improve coordination.