Add frame functionality to the game
I have implemented multiple frames in the Window class to facilitate the
display of relevant information as the game progresses.

- The first frame serves as the launch stage of the game, where the player
can press a button to initiate the game.

- The second frame represents the main game loop, where both the human
and computer players navigate the maze.

-The last frame is dedicated to displaying the game results upon completion.
Separate classes were created for each frame, and the canvas along with
its associated methods were moved to the GameFrame class.

Additionally, I have disabled the resizing of the game window for
consistency.

These modifications enhance the organization and user experience of the
game.